% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nets.R
\name{load_nets}
\alias{load_nets}
\title{Load chapter datasets into the workspace}
\usage{
load_nets(filename)
}
\arguments{
\item{filename}{Name of the .RData file in extdata}
}
\description{
Load chapter datasets into the workspace
}
